修改了一些错误

+++ src/_10_binarysystem/NumberOf1.java
@@ -2,8 +2,8 @@ package _10_binarysystem;
 
 //输出整数转化为二进制后1的个数
 public class NumberOf1 {
-	public static int numOf1(int n) {
-		int count = 0;
+	public static long numOf1(long n) {
+		long count = 0;
 		while (n != 0) {
 			count++;
 			n = (n - 1) & n;
